MigrationPython/requirements.txt

azure-cosmos
azure-identity
InquirerPy
dotenv
python-dateutil